Record feed with failed loading image
