+++ EssentialFeediOS/Feed UI/Views/FeedImageCell.swift
@@ -111,6 +111,7 @@ public final class FeedImageCell: UITableViewCell {
         let button = UIButton()
         button.translatesAutoresizingMaskIntoConstraints = false
         button.setTitle("↻", for: .normal)
+        button.backgroundColor = .lightGray
         button.titleLabel?.textColor = .white
         button.titleLabel?.font = .systemFont(ofSize: 60)
         button.addTarget(self, action: #selector(retryButtonTapped), for: .touchUpInside)
@@ -136,31 +137,4 @@ public final class FeedImageCell: UITableViewCell {
         label.adjustsFontForContentSizeCategory = true
         return label
     }()
-}
-
-extension UIView {
-    func pinToSuperview(constant: CGFloat = 0) {
-        guard let superview = superview else { return }
-
-        NSLayoutConstraint.activate([
-            self.topAnchor.constraint(equalTo: superview.topAnchor, constant: constant),
-            self.leadingAnchor.constraint(equalTo: superview.leadingAnchor, constant: constant),
-            self.trailingAnchor.constraint(equalTo: superview.trailingAnchor, constant: -constant),
-            self.bottomAnchor.constraint(equalTo: superview.bottomAnchor, constant: -constant)
-        ])
-    }
-
-    func pinHorizontally(to view: UIView) {
-        NSLayoutConstraint.activate([
-            self.leadingAnchor.constraint(equalTo: view.leadingAnchor),
-            self.trailingAnchor.constraint(equalTo: view.trailingAnchor)
-        ])
-    }
-
-    func pinVertically(to view: UIView) {
-        NSLayoutConstraint.activate([
-            self.topAnchor.constraint(equalTo: view.topAnchor),
-            self.bottomAnchor.constraint(equalTo: view.bottomAnchor)
-        ])
-    }
 }
